requirements.bot.txt
aiogram
python-dotenv
requests
llama-cpp-python
openai-whisper
faster-whisper
numpy
scipy
faiss-cpu
sentence-transformers
openai  # fallback
sqlalchemy
pymysql
sentence-transformers
faiss-cpu
gpt4all
scikit-learn
mysqlclient
ffmpeg-python
aiohttp
